refactor(node): extract runtime orchestration module (#3604)
